requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
python-dotenv==1.0.0
cryptography==41.0.7
gunicorn==21.2.0
pytest==7.4.3
